Fix iPad photo orientation
On iPad the photos from the QML ImageCapture imageSaved signal do not have the
correct orientation. We could try to work out the orientation from the image
metadata or try to grab the device orientation at the point the photo is taken.
The preview image appears to have the correct orientation and same resolution
as the saved image. So I've changed the CameraQml class to use this instead.