Use our own custom Qt build for running Qt tests
Previously we were using a third party distribution of Qt for running tests. This is fine
for some tests but it is useful to have the other third party dependencies for e.g.
SQL Cipher database tests. It's also more likely to flag problems with our Qt etc builds.
While GitHub supports uploading and downloading of artifacts that can be shared between jobs
they expire after 90 days. So instead I've (ab)used the release mechanism so that Qt is
built as an artifact of a draft CamCOPS release (along with the other third party components).
We'll need to clean up old builds manually.